Allow changes to game pending update status
- Added the ability to set a game as updated, with the intention of it being removed from Steams scheduled update queue.

Unfortunately this doesn't work in practise. Steam automatically rechecks for updates on startup and will revert all changes I make to the .acf files.

This is disappointing and I haven't found a work-around as of yet.

I intend to keep digging to see if there is an elegant solution to this.